backend/requirements/requirements.txt
Django==1.8.16
pymongo==3.4.0
django-crispy-forms==1.6.0
django-pagedown==0.1.1
djangorestframework-jwt==1.8.0
